requirements.txt
aiofiles
python-dotenv
fastapi[all]
uvicorn
pipecat-ai[daily,websocket]